Mascot + UI update
Sorry about not putting a description on the last commit. Anyway, I've added areas for all the UI doohickeys we're going to be implementing, and tweaked the CSS to place them on the sides of the viewing pane.

(I can change the flex behavior back if we decide it looked better the other way.)